Move `get_hash` to `strutil`
