+++ src/python/pants/util/strutil_test.py
@@ -2,10 +2,12 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 import textwrap
+from dataclasses import dataclass
 from textwrap import dedent
 
 import pytest
 
+from pants.util.frozendict import FrozenDict
 from pants.util.strutil import (
     bullet_list,
     comma_separated_list,
@@ -14,6 +16,7 @@ from pants.util.strutil import (
     ensure_text,
     first_paragraph,
     fmt_memory_size,
+    get_stable_hash,
     hard_wrap,
     path_safe,
     pluralize,
@@ -397,3 +400,19 @@ def test_docstring_decorator() -> None:
 
     with pytest.raises(AssertionError):
         assert show_why_this_is_needed.__doc__ == "calc 1 + 1 = 2"
+
+
+def test_hash() -> None:
+    @dataclass(frozen=True)
+    class Data:
+        mapping: FrozenDict[str, str]
+
+    data = Data(
+        FrozenDict(
+            {alpha: alpha.lower() for alpha in [chr(a) for a in range(ord("A"), ord("Z") + 1)]}
+        )
+    )
+    assert (
+        get_stable_hash(data).hexdigest()
+        == "e4da3c55de6ce98ddcbd5b854ff01f5c8b47fdcb2e10ddd5176505e39a332730"
+    )
